Added new source code files
I added: crc.c, crc.h,Pressure_Comp., and Pressure_Comp.h. These source files are to be used for pressure compensation and is to be called from the main function in the main.cpp file. These have not been fully implemented and connected for an accurate pressure compensated reading.